test: bc layer for deprecated static::$container
