+++ tests/BcLayerWebTestCase.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Mtarld\ApiPlatformMsBundle\Tests;
+
+use LogicException;
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
+use Symfony\Component\HttpKernel\KernelInterface;
+
+// Remove after support of symfony support of 4.4 is dropped
+abstract class BcLayerWebTestCase extends WebTestCase
+{
+    private static $kernelContainer;
+
+    protected static function bootKernel(array $options = []): KernelInterface
+    {
+        $kernel = parent::bootKernel($options);
+
+        self::$kernelContainer = $kernel->getContainer();
+
+        return $kernel;
+    }
+
+    protected static function getContainer(): ContainerInterface
+    {
+        if (method_exists(parent::class, 'getContainer')) {
+            return parent::getContainer();
+        }
+
+        if (!static::$booted) {
+            static::bootKernel();
+        }
+
+        try {
+            return self::$kernelContainer->get('test.service_container');
+        } catch (ServiceNotFoundException $e) {
+            throw new LogicException('Could not find service "test.service_container". Try updating the "framework.test" config to "true".', 0, $e);
+        }
+    }
+}
