+++ tests/Microservice/MicroservicePoolTest.php
@@ -6,7 +6,7 @@ use Mtarld\ApiPlatformMsBundle\Exception\MicroserviceConfigurationException;
 use Mtarld\ApiPlatformMsBundle\Exception\MicroserviceNotConfiguredException;
 use Mtarld\ApiPlatformMsBundle\Microservice\Microservice;
 use Mtarld\ApiPlatformMsBundle\Microservice\MicroservicePool;
-use Mtarld\ApiPlatformMsBundle\Tests\BcLayer\BcLayerKernelTestCase;
+use Mtarld\ApiPlatformMsBundle\Tests\BcLayerKernelTestCase;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
