+++ tests/Serializer/ObjectDenormalizerTest.php
@@ -2,7 +2,7 @@
 
 namespace Mtarld\ApiPlatformMsBundle\Tests\Serializer;
 
-use Mtarld\ApiPlatformMsBundle\Tests\BcLayer\BcLayerKernelTestCase;
+use Mtarld\ApiPlatformMsBundle\Tests\BcLayerKernelTestCase;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Dto\PuppyDto;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Entity\Puppy;
 use Symfony\Component\Serializer\SerializerInterface;
